Made more explicit that the dY/dt criteria to stop the algorithm is based on computing the maximum scaled Euclidean norm rather than using a distance. Thus we moved the compute_norms function to the nebm_base class and commented the redefined method in the Geodesic code. Also added a filter to avoid using latice/mesh sites without material. Using the distance should be equivalent but the stopping criteria values would change, we could try this in the future
